Add reusable PySide canvas widget
